refactor: Move time-range-options from constants/ to utils/
The module now contains functions (isValidLookback, lookbackFromDuration)
in addition to constants, so utils/ is a more appropriate home.
