+++ packages/jaeger-ui/src/utils/time-range-options.ts
